Start to get real Cash extractions to work

+++ ethereum/contracts/test/MockCashToken.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.1;
 import "../CashToken.sol";
 
 contract MockCashToken is CashToken {
-	constructor(address admin_, uint initialSupply_, address holder_) CashToken(admin_) {
+	constructor(address admin_, uint128 initialSupply_, address holder_) CashToken(admin_) {
 		cashPrincipal[holder_] = initialSupply_;
 		totalCashPrincipal = initialSupply_;
 	}
